OMAP: USB Host: Fix runtime suspend
ehci_omap_bus_suspend() is called at runtime suspend, when the
device is still running but the bus is idle.  This routine used
to call pm_runtime_put_sync() which resulted in clocks to usbtll
being turned off.  This meant that we could not resume the bus when
a device was plugged into the USB host port.  The driver now uses
driver suspend/resume routines which means we don't do this until
the device (not just the bus) enters suspend.

Also add support for turning off/on a clock connected to the USB
PHY, such as auxclk3, at suspend/resume.

Signed-off-by: Steve Schefter <steve@scheftech.com>